Better split between business logic & UI code for EIP-712
