--- src_bagl/ui_flow_712.h
@@ -1,12 +1,0 @@
-#ifndef UI_FLOW_712_H_
-#define UI_FLOW_712_H_
-
-#ifdef HAVE_EIP712_FULL_SUPPORT
-
-#include "ux_flow_engine.h"
-
-extern const ux_flow_step_t* const ux_712_flow[];
-
-#endif  // HAVE_EIP712_FULL_SUPPORT
-
-#endif  // UI_FLOW_712_H_
